Moved generic functionality into lib
