+++ app/models/api/v3/node.rb
@@ -110,47 +110,6 @@ module Api
         rel = rel.where('node_inds.year' => year) if year.present?
         rel
       end
-
-      def flow_values(context, year, attribute)
-        flow_values_for_attributes(context, year, [attribute])
-      end
-
-      def flow_values_for_attributes(context, year, attributes)
-        node_index = NodeType.node_index_for_id(context, node_type_id)
-        attributes_ids = attributes.map(&:id)
-        attribute_type = attributes.first&.class&.name&.demodulize&.downcase
-        flow_values_table = :"flow_#{attribute_type}s"
-        Flow.
-          joins(flow_values_table).
-          where("#{flow_values_table}.#{attribute_type}_id" => attributes_ids).
-          where('path[?] = ?', node_index, id).
-          where(context_id: context.id, year: year)
-      end
-
-      def flow_values_totals_for_attributes_into(context, year, attributes, other_node_type, other_node_id)
-        other_node_index = NodeType.node_index_for_name(context, other_node_type)
-        attribute_type = attributes.first&.class&.name&.demodulize&.downcase
-        flow_values_table = :"flow_#{attribute_type}s"
-        attributes_table = :"#{attribute_type}s"
-        nodes_join_clause = ActiveRecord::Base.send(
-          :sanitize_sql_array,
-          ['JOIN nodes ON nodes.id = flows.path[?]',
-           other_node_index]
-        )
-        group_clause = ActiveRecord::Base.send(
-          :sanitize_sql_array,
-          ['flows.path[?], quants.name',
-           other_node_index]
-        )
-
-        flow_values_for_attributes(context, year, attributes).
-          select("SUM(CAST(#{flow_values_table}.value AS DOUBLE PRECISION)) AS value, #{attributes_table}.name").
-          joins(flow_values_table => attribute_type).
-          joins(nodes_join_clause).
-          where('? = path[?]', other_node_id, other_node_index).
-          where('NOT is_unknown').
-          group(group_clause)
-      end
     end
   end
 end
