+++ app/services/api/v3/place_node/indicators_table.rb
@@ -2,6 +2,8 @@ module Api
   module V3
     module PlaceNode
       class IndicatorsTable
+        include Api::V3::Profiles::AttributesInitializer
+
         def initialize(context, year, node)
           @context = context
           @year = year
@@ -140,35 +142,6 @@ module Api
               }
             ]
           }
-        end
-
-        def initialize_attributes(attributes_list)
-          attributes = attributes_list.map do |attribute_hash|
-            attribute_hash.merge(
-              attribute: initialize_attribute_from_hash(
-                attribute_hash
-              )
-            )
-          end
-          attributes.select do |attribute_hash|
-            attribute_hash && attribute_hash[:attribute].present?
-          end
-        end
-
-        def initialize_attribute_from_hash(attribute_hash)
-          dictionary =
-            if attribute_hash[:attribute_type] == 'quant'
-              Dictionary::Quant.instance
-            elsif attribute_hash[:attribute_type] == 'ind'
-              Dictionary::Ind.instance
-            end
-          return nil unless dictionary
-          attribute = dictionary.get(attribute_hash[:attribute_name])
-          if attribute.nil?
-            Rails.logger.debug 'NOT FOUND ' + attribute_hash[:attribute_name]
-          end
-
-          attribute
         end
       end
     end
