+++ lib/api/v3/profiles/top_nodes_list.rb
@@ -1,6 +1,6 @@
 module Api
   module V3
-    module PlaceNode
+    module Profiles
       class TopNodesList
         def initialize(context, year, node, data)
           @context = context
